// frontend/src/app/models/CustomizeOrderModel.ts
export interface CustomizeOrderModel {
    customizedServiceId?: number;
    userName?: string;
    userMobile?: string;
    projectName?: string;
    unitNumber?: string;
    targetDate?: Date;
    createdDate?: Date;
    canceledDate?: null;
    deliveredDate?: Date;
    assignedDate?: Date;
    price?: number;
    status?: string;
    rated?: boolean;
    rate?: string;
    comment?: string;
    serviceTypeName?: string;
    technician?: string;
    assignedBy?: string;
    closedBy?: string;
    photo?: string;
    ownerName?: string;
    description?: string;
    userId?: string;
}